pyproject.toml
[project]
name = "airflow-jaffleshop"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "apache-airflow>=3.1.7",
    "dbt-core>=1.11.4",
    "dbt-snowflake>=1.11.2",
]
